Moved findCommonParent tests to highlight.cy.ts
