+++ test/cypress/e2e/highlight.cy.ts
@@ -1,4 +1,4 @@
-import { getUnionDOMRect } from "cumulocity-cypress/lib/commands/highlight";
+import { findCommonParent, getUnionDOMRect } from "cumulocity-cypress/lib/commands/highlight";
 
 describe("highlight", () => {
   beforeEach(() => {
@@ -231,6 +231,27 @@ describe("highlight", () => {
             y: 0,
           });
         });
+      });
+    });
+  });
+
+  describe("findCommonParent", () => {
+    it("should return the common parent element 1", () => {
+      cy.visit("/dtm.newasset.html").then(() => {
+        const $elements = Cypress.$(".bottom-drawer .card-footer button");
+        const commonParent = findCommonParent($elements);
+        expect(commonParent).to.have.class("card-footer");
+      });
+    });
+
+    it("should return the common parent element 2", () => {
+      cy.visit("/dtm.newasset.html").then(() => {
+        const $elements = Cypress.$("main .d-flex");
+        expect($elements.length).to.eq(8);
+        const commonParent = findCommonParent($elements);
+        expect(commonParent?.tagName?.toLocaleLowerCase()).to.eq(
+          "c8y-repeat-section"
+        );
       });
     });
   });
